Consolidate nested classes.
