--- osu.Game/Graphics/UserInterface/OsuDropdownHeader.cs
@@ -1,69 +1,0 @@
-// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
-// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
-
-using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Primitives;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.UserInterface;
-using osu.Game.Graphics.Sprites;
-using OpenTK.Graphics;
-using osu.Framework.Extensions.Color4Extensions;
-
-namespace osu.Game.Graphics.UserInterface
-{
-    public class OsuDropdownHeader : DropdownHeader
-    {
-        private SpriteText label;
-        protected override string Label
-        {
-            get { return label.Text; }
-            set { label.Text = value; }
-        }
-
-        private Color4? accentColour;
-        public virtual Color4 AccentColour
-        {
-            get { return accentColour.GetValueOrDefault(); }
-            set
-            {
-                accentColour = value;
-                BackgroundColourHover = value;
-            }
-        }
-
-        public OsuDropdownHeader()
-        {
-            Foreground.Padding = new MarginPadding(4);
-
-            AutoSizeAxes = Axes.None;
-            Margin = new MarginPadding { Bottom = 4 };
-            CornerRadius = 4;
-            Height = 40;
-
-            Foreground.Children = new Drawable[]
-            {
-                label = new OsuSpriteText
-                {
-                    Anchor = Anchor.CentreLeft,
-                    Origin = Anchor.CentreLeft,
-                },
-                new TextAwesome
-                {
-                    Icon = FontAwesome.fa_chevron_down,
-                    Anchor = Anchor.CentreRight,
-                    Origin = Anchor.CentreRight,
-                    Margin = new MarginPadding { Right = 4 },
-                    TextSize = 20
-                }
-            };
-        }
-
-        [BackgroundDependencyLoader]
-        private void load(OsuColour colours)
-        {
-            BackgroundColour = Color4.Black.Opacity(0.5f);
-            BackgroundColourHover = accentColour ?? colours.PinkDarker;
-        }
-    }
-}
